feature: Asserts table resizable and orderable
