+++ packages/bruno-app/src/components/RequestPane/Assertions/StyledWrapper.js
@@ -20,14 +20,6 @@ const Wrapper = styled.div`
     td {
       padding: 6px 10px;
 
-      &:nth-child(1) {
-        width: 30%;
-      }
-
-      &:nth-child(4) {
-        width: 70px;
-      }
-
       select {
         background-color: transparent;
       }
